+++ package.json
@@ -7,7 +7,9 @@
     "serve": "vite serve"
   },
   "dependencies": {
-    "vue": "^3.0.5"
+    "vue": "^3.0.11",
+    "vue-router": "^4.0.6",
+    "vuetify": "^2.4.11"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^1.2.1",
