+++ src/components/landingComponent.vue
@@ -1,0 +1,19 @@
+<template>
+  <h1>{{ msg }}</h1>
+</template>
+
+<script setup>
+import { defineProps, reactive } from 'vue'
+
+defineProps({
+  msg: String
+})
+
+const state = reactive({ count: 0 })
+</script>
+
+<style scoped>
+a {
+  color: #42b983;
+}
+</style>
